test(e2e): unify slashing timing profiles and merge duplicate suites
Round-2 consolidation of multi-node/slashing:

- Add a SENTINEL_TIMING profile to setup.ts and route every test file's slot
  timing through it or baseSlashingOpts, removing the per-file redeclared
  ETHEREUM_SLOT_DURATION/AZTEC_SLOT_DURATION/AZTEC_EPOCH_DURATION constants.
- Merge duplicate_attestation + duplicate_proposal into equivocation_offenses
  with a shared scenario runner and two describes.
- Fold inactivity_slash_with_consecutive_epochs into inactivity_slash (2 describes).
- Centralize the findSlashOffense offense lookup (and SlashOffense type) in setup.ts.
- Remove the dead shouldVeto:false branch in slash_veto_demo.
- Hoist sentinel_status_slash its 1-2 onto a shared runProposerFaultScenario helper.
