+++ yarn-project/end-to-end/src/multi-node/slashing/setup.ts
@@ -8,7 +8,7 @@ import { unique } from '@aztec/foundation/collection';
 import { EthAddress } from '@aztec/foundation/eth-address';
 import { retryUntil } from '@aztec/foundation/retry';
 import { pluralize } from '@aztec/foundation/string';
-import { getRoundForOffense } from '@aztec/slasher';
+import { type OffenseType, getRoundForOffense } from '@aztec/slasher';
 import type { AztecNodeAdmin } from '@aztec/stdlib/interfaces/client';
 import type { TxHash } from '@aztec/stdlib/tx';
 
@@ -59,6 +59,42 @@ export const baseSlashingOpts = {
   slashDuplicateProposalPenalty: slashingUnit,
   slashingOffsetInRounds: 1,
 };
+
+/**
+ * The fast slot timing shared by the sentinel-observation suites (`sentinel_status_slash`,
+ * `validators_sentinel`, `multiple_validators_sentinel`, `slash_veto_demo`). Slots are 8s (eth 4s ×
+ * epoch 2), which fit a checkpoint comfortably at the fast mocked-p2p operational budgets used at
+ * eth<8, keeping wall-clock down for tests whose bodies advance through real L2 slots. Spread into a
+ * {@link MultiNodeTestContext.setup} call alongside {@link SLASHER_ENABLED_MULTI_VALIDATOR_OPTS},
+ * `initialValidators`, and the per-test slash/penalty config. Non-sentinel offense-detection files
+ * that only need the same fast timing spread this and override `sentinelEnabled: false`.
+ */
+export const SENTINEL_TIMING = {
+  anvilSlotsInAnEpoch: 4,
+  listenAddress: '127.0.0.1',
+  ethereumSlotDuration: 4,
+  aztecSlotDuration: 8,
+  aztecEpochDuration: 2,
+  sentinelEnabled: true,
+} as const;
+
+/** A single detected slash offense as returned by {@link AztecNodeService.getSlashOffenses}. */
+export type SlashOffense = Awaited<ReturnType<AztecNodeService['getSlashOffenses']>>[number];
+
+/** Looks up the offense recorded for `validator` of `offenseType` at `slot`, if any. */
+export function findSlashOffense(
+  offenses: SlashOffense[],
+  validator: EthAddress,
+  offenseType: OffenseType,
+  slot: SlotNumber,
+): SlashOffense | undefined {
+  return offenses.find(
+    offense =>
+      offense.validator.equals(validator) &&
+      offense.offenseType === offenseType &&
+      offense.epochOrSlot === BigInt(slot),
+  );
+}
 
 export function awaitProposalExecution(
   slashingProposer: SlashingProposerContract,
